added Cart html panels to renderIntoPanels and Modularized
Modularized getting the name from the HTML and added function to get the name for the shopping cart case. Now iterates through all the html panels it finds (both shopping cart and enrolled classes in the cart page or all panels in search page). NEED TO FIX: formatting and when click on the info icon page reloads